Major update on querying and multi-treading
I know that this is in the new heatMap branch but this is a major breakthrough with querying the data. BarChart is now responsive to different flight carriers and for no restriction.
TO-DO:
Add some more labels, for barChart and heatMap class to make it all look neat and together